Introduced R<> to pass references and T*Result<>. Significant rework covering DataSources, removing FileReaders, and migrating *Parcellers to *Readers. Migrated CImage -> Bitmap from I<CDataSource> to CData. Updated CData types to UInt64 from UInt32. Introduced SValue::mEmpty to represent "null" values.

+++ Source/Base/CData.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "CString.h"
+#include "SError.h"
 
 //----------------------------------------------------------------------------------------------------------------------
 // MARK: CData
@@ -13,8 +13,8 @@ class CDataInternals;
 class CData {
 	// Types
 	public:
-		typedef	UInt32	Size;
-		typedef	UInt32	ByteIndex;
+		typedef	UInt64	Size;
+		typedef	UInt64	ByteIndex;
 
 	// Methods
 	public:
